requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.1
python-telegram-bot==12.6.1
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
git+https://github.com/rafaelassacconi/stackoverflow.git